made the suggested changes to the SignUp component
